Refactoring (dom module)
